requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
bcrypt==4.0.1
